alx_project_nexus/requirements.txt
aiosqlite==0.21.0
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
executing==2.2.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
mysql-connector-python==8.0.28
packaging==25.0
parso==0.8.4
promise==2.3
prompt_toolkit==3.0.51
protobuf==6.31.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
